requirements.txt
# Framework
# Django==1.6
http://www.djangoproject.com/m/releases/1.6/Django-1.6b1.tar.gz

# Libs
pep8==1.4.6

# Apps
django-pipeline==1.3.14
django-jenkins==0.14.0
django-bower==5.0.1
